alxET_rt_api/requirements.txt
django>=5.2.5
django-cors-headers>=4.7.0
django-redis>=6.0.0
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
drf-spectacular>=0.28.0
google-api-python-client>=2.179.0
google-auth>=2.40.3
gspread>=6.2.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.1.1
redis[hiredis]>=6.4.0
whitenoise>=6.9.0
celery>=5.5.3
requests>=2.31.0
openpyxl>=3.1.5
flower

# Testing packages 
# pytest 
# pytest-django 
# pytest-factoryboy 
# coverage 
# pytest-cov 
# pytest-html
# locust